move protocol test to protocol test directory (#3864)
1. move the protocol test to right directory
2. fix wrong string format params count
