+++ protocol/src/test/java/com/alibaba/otter/canal/protocol/SecurityUtilTest.java
@@ -1,4 +1,4 @@
-package com.alibaba.otter.canal.server;
+package com.alibaba.otter.canal.protocol;
 
 import java.security.NoSuchAlgorithmException;
